Point two doc guards at the relocated refresh module
#518 moved services/refresh.py to orchestration/refresh.py and left two
test imports on the old path. Both fail with ModuleNotFoundError on main's
own CI run at 3d698a7a, so this repairs them ahead of the work that
follows. A third site in one of the same files was already updated.
